Setup combat XP map - Use float in more places

+++ src/main/java/com/gmail/nossr50/config/hocon/experience/ConfigExperienceArchery.java
@@ -6,7 +6,7 @@ import ninja.leaping.configurate.objectmapping.serialize.ConfigSerializable;
 @ConfigSerializable
 public class ConfigExperienceArchery {
 
-    public static final double DISTANCE_MULTIPLIER_DEFAULT = 0.025D;
+    public static final float DISTANCE_MULTIPLIER_DEFAULT = 0.025F;
 
     @Setting(value = "Distance-Multiplier", comment = "The distance multiplier is multiplied against the distance an " +
             "arrow travels before hitting its target to determine final XP values awarded." +
@@ -14,9 +14,9 @@ public class ConfigExperienceArchery {
             "\nDistance is in blocks traveled." +
             "\nThis value is added on to normal XP gains from damage for Archery." +
             "\nDefault value: " + DISTANCE_MULTIPLIER_DEFAULT)
-    private double distanceMultiplier = DISTANCE_MULTIPLIER_DEFAULT;
+    private float distanceMultiplier = DISTANCE_MULTIPLIER_DEFAULT;
 
-    public double getDistanceMultiplier() {
+    public float getDistanceMultiplier() {
         return distanceMultiplier;
     }
 }
